+++ config/SOUE01/symbols.txt
@@ -919,9 +919,9 @@ fn_80027490 = .text:0x80027490; // type:function size:0x40
 fn_800274D0 = .text:0x800274D0; // type:function size:0x40
 fn_80027510 = .text:0x80027510; // type:function size:0x4C
 fn_80027560 = .text:0x80027560; // type:function size:0x5C
-fn_800275C0 = .text:0x800275C0; // type:function size:0x44
+__ct__13EffectsStructFv = .text:0x800275C0; // type:function size:0x44
 fn_80027610 = .text:0x80027610; // type:function size:0x44
-fn_80027660 = .text:0x80027660; // type:function size:0x78
+__dt__13EffectsStructFv = .text:0x80027660; // type:function size:0x78
 fn_800276E0 = .text:0x800276E0; // type:function size:0x84
 fn_80027770 = .text:0x80027770; // type:function size:0xE8
 fn_80027860 = .text:0x80027860; // type:function size:0xC8
@@ -3346,12 +3346,12 @@ __dt__14InteractionMdlFv = .text:0x80093E90; // type:function size:0xA8
 __dt__16AttentionManagerFv = .text:0x80093F40; // type:function size:0xCC
 create__16AttentionManagerFv = .text:0x80094010; // type:function size:0x19C
 createHeap__16AttentionManagerFv = .text:0x800941B0; // type:function size:0x204
-isInNormalGameState__16AttentionManagerFv = .text:0x800943C0; // type:function size:0x98
-isZButtonPressed__16AttentionManagerFv = .text:0x80094460; // type:function size:0x4
-isZButtonHeld__16AttentionManagerFv = .text:0x80094470; // type:function size:0x4
-checkUnknown__16AttentionManagerFv = .text:0x80094480; // type:function size:0x8
-checkLink1__16AttentionManagerFv = .text:0x80094490; // type:function size:0x4C
-checkLink2__16AttentionManagerFv = .text:0x800944E0; // type:function size:0xB4
+isInNormalGameState__16AttentionManagerCFv = .text:0x800943C0; // type:function size:0x98
+isZButtonPressed__16AttentionManagerCFv = .text:0x80094460; // type:function size:0x4
+isZButtonHeld__16AttentionManagerCFv = .text:0x80094470; // type:function size:0x4
+checkUnknown__16AttentionManagerCFv = .text:0x80094480; // type:function size:0x8
+checkLink1__16AttentionManagerCFv = .text:0x80094490; // type:function size:0x4C
+checkLink2__16AttentionManagerCFv = .text:0x800944E0; // type:function size:0xB4
 execute__16AttentionManagerFv = .text:0x800945A0; // type:function size:0x284
 switchTarget__16AttentionManagerFl = .text:0x80094830; // type:function size:0xCA4
 debugPrint8__FPCce = .text:0x800954E0; // type:function size:0x50
@@ -3367,7 +3367,7 @@ fn_800964B0__14AttentionGroupFv = .text:0x800964B0; // type:function size:0xBC
 fn_80096570 = .text:0x80096570; // type:function size:0x10C
 fn_80096680 = .text:0x80096680; // type:function size:0x3DC
 fn_80096A60 = .text:0x80096A60; // type:function size:0x50
-fn_80096AB0 = .text:0x80096AB0; // type:function size:0x88
+targetScore__16AttentionManagerFP12dAcObjBase_cP12dAcObjBase_c = .text:0x80096AB0; // type:function size:0x88
 fn_80096B40__16AttentionManagerFP12dAcObjBase_c = .text:0x80096B40; // type:function size:0x80
 addPickUpTarget__16AttentionManagerFR12dAcObjBase_cf = .text:0x80096BC0; // type:function size:0x18
 addSitTarget__16AttentionManagerFR12dAcObjBase_cUlf = .text:0x80096BE0; // type:function size:0x20
@@ -17274,7 +17274,7 @@ fn_802E0190 = .text:0x802E0190; // type:function size:0x13C
 fn_802E02D0 = .text:0x802E02D0; // type:function size:0xF8
 fn_802E03D0 = .text:0x802E03D0; // type:function size:0x174
 fn_802E0550 = .text:0x802E0550; // type:function size:0x124
-fn_802E0680 = .text:0x802E0680; // type:function size:0x1C
+targetAngleY__4cLibFRC7mVec3_cRC7mVec3_c = .text:0x802E0680; // type:function size:0x1C
 fn_802E06A0 = .text:0x802E06A0; // type:function size:0x68
 fn_802E0710 = .text:0x802E0710; // type:function size:0xD8
 fn_802E07F0 = .text:0x802E07F0; // type:function size:0x24
