+++ include/c/c_lib.h
@@ -1,0 +1,13 @@
+#ifndef C_CLIB_H
+#define C_CLIB_H
+
+#include <common.h>
+#include <m/m_vec.h>
+
+namespace cLib {
+
+s32 targetAngleY(const mVec3_c& target, const mVec3_c& source);
+
+} // namespace cLib
+
+#endif
